+++ CHANGELOG.md
@@ -11,4 +11,4 @@
 - Provided prebuilt [Win64 Brotli native shared library](lib/src/brotli/blobs/esbrotli-win64.dll)
 - Provided prebuilt [Win64 LZ4 native shared library](lib/src/lz4/blobs/eslz4-win64.dll)
 - Provided prebuilt [Win64 Zstd (ZStandard) native shared library](lib/src/zstd/blobs/eszstd-win64.dll)
-- Implemented flexible library loading with [OpenLibrary](lib/src/common/ffi/openlibrary.dart) mixin
+- Implemented flexible library loading with [OpenLibrary](lib/src/framework/native/openlibrary.dart) mixin